rename initial_answer_generation
